refactor: Replace Serializer.Anything with Serializer for consistency and clarity in serialization methods
