+++ Assets/root/Editor/Scripts/UI/Window/MainWindowEditor.cs
@@ -36,11 +36,7 @@ namespace com.IvanMurzak.Unity.MCP.Editor
             EditorUtility.SetDirty(McpPluginUnity.AssetFile); // Undo record completed
         }
 
-        private void OnChanged(McpPluginUnity.Data data)
-        {
-            Debug.Log($"OnChanged");
-            Repaint();
-        }
+        private void OnChanged(McpPluginUnity.Data data) => Repaint();
 
         private void OnEnable()
         {
