+++ Assets/root/Runtime/Reflection/Custom/RS_UnityEngineObject.cs
@@ -25,8 +25,8 @@ namespace com.IvanMurzak.Unity.MCP.Utils
                     {
                         name = name,
                         type = type.FullName,
-                        fields = Serializer.Anything.SerializeFields(obj, flags),
-                        properties = Serializer.Anything.SerializeProperties(obj, flags)
+                        fields = Serializer.SerializeFields(obj, flags),
+                        properties = Serializer.SerializeProperties(obj, flags)
                     }.SetValue(new InstanceID(unityObject.GetInstanceID()));
                 }
                 else
