--- Assets/root/Runtime/Reflection/ReflectionUtils.Unity.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a144723f1c4435d4882be54644e07fcb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
