--- Assets/root/Runtime/Reflection/Serializer.Anything.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bc8089c9951fbc4da44f99efaca2de7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
